+++ src/main/java/kokopu/com/Kokopu/service/EmployeeDetail.java
@@ -8,6 +8,6 @@ import kokopu.com.Kokopu.model.Employee;
 @Service
 public interface EmployeeDetail {
 	
-	public Employee getEmployeeDetails(int empId);
+	public Employee getEmployeeDetails(long empId);
 
 }
